+++ examples/hello_nvme_bdev_rust_wrapper/README.md
@@ -1,4 +1,4 @@
 This directory contains Rust version of hello_nvme_bdev.c program. Unlike the program
 inside hello_nvme_bdev_rust, the program here uses the Rust API we developed from
-rust-spdk. This program is used as a way to test the functionality of Rust API we 
+spdk-rs. This program is used as a way to test the functionality of Rust API we 
 designed.
