+++ spdk-rs/src/env.rs
@@ -1,0 +1,9 @@
+/*************************************************************************
+  > File Name:       env.rs
+  > Author:          Zeyuan Hu
+  > Mail:            iamzeyuanhu@utexas.edu
+  > Created Time:    10/10/18
+  > Description:
+    
+    FFI for "env.h"
+ ************************************************************************/
