+++ spdk-rs/src/lib.rs
@@ -7,6 +7,8 @@ extern crate libc;
 
 mod event;
 mod bdev;
+mod context;
 
-pub use event::{AppOpts, AppContext, app_stop};
+pub use event::{SpdkAppOpts, spdk_app_stop};
 pub use bdev::{Bdev, BdevDesc};
+pub use context::{AppContext};
